# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistema de Gerenciamento de Classificação Indicativa — Portaria 368/2014 MJ
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /conteudos:
    get:
      operationId: listConteudos
      tags: [conteudos]
      summary: Listar conteúdos
      parameters:
        - name: tipo
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: busca
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de conteúdos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conteudo"
    post:
      operationId: createConteudo
      tags: [conteudos]
      summary: Cadastrar conteúdo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConteudoInput"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conteudo"
  /conteudos/{id}:
    get:
      operationId: getConteudo
      tags: [conteudos]
      summary: Obter conteúdo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conteúdo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conteudo"
    patch:
      operationId: updateConteudo
      tags: [conteudos]
      summary: Atualizar conteúdo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConteudoUpdate"
      responses:
        "200":
          description: Atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conteudo"
    delete:
      operationId: deleteConteudo
      tags: [conteudos]
      summary: Remover conteúdo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removido

  # ─── Solicitações ─────────────────────────────────────────────────────────
...
  /classificacoes:
    get:
      operationId: listClassificacoes
      tags: [classificacoes]
      summary: Listar classificações emitidas
      parameters:
        - name: faixa
          in: query
          schema:
            type: string
        - name: tipo_conteudo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de classificações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Classificacao"
  /classificacoes/{id}:
    get:
      operationId: getClassificacao
      tags: [classificacoes]
      summary: Obter classificação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Classificação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classificacao"
    patch:
      operationId: updateClassificacao
      tags: [classificacoes]
      summary: Atualizar classificação (revisão)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassificacaoUpdate"
      responses:
        "200":
          description: Atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classificacao"

  # ─── Descritores ──────────────────────────────────────────────────────────
  /descritores:
    get:
      operationId: listDescritores
      tags: [descritores]
      summary: Listar descritores de conteúdo
      responses:
        "200":
          description: Lista de descritores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Descritor"
    post:
      operationId: createDescritor
      tags: [descritores]
      summary: Criar descritor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DescritorInput"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Descritor"
  /descritores/{id}:
    patch:
      operationId: updateDescritor
      tags: [descritores]
      summary: Atualizar descritor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DescritorUpdate"
      responses:
        "200":
          description: Atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Descritor"
    delete:
      operationId: deleteDescritor
      tags: [descritores]
      summary: Remover descritor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removido

  # ─── Dashboard ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Conteudo:
      type: object
      required: [id, titulo, tipo, ano_producao, pais_origem, status, criado_em]
      properties:
        id:
          type: integer
        titulo:
          type: string
        titulo_original:
          type: ["string", "null"]
        tipo:
          type: string
          enum: [filme, serie, documentario, animacao, jogo, publicidade, outros]
        ano_producao:
          type: integer
        pais_origem:
          type: string
        distribuidora:
          type: ["string", "null"]
        sinopse:
          type: ["string", "null"]
        duracao_min:
          type: ["integer", "null"]
        status:
          type: string
          enum: [aguardando, em_analise, classificado, rejeitado]
        criado_em:
          type: string
          format: date-time
        atualizado_em:
          type: string
          format: date-time
    ConteudoInput:
      type: object
      required: [titulo, tipo, ano_producao, pais_origem]
      properties:
        titulo:
          type: string
          minLength: 1
        titulo_original:
          type: string
        tipo:
          type: string
          enum: [filme, serie, documentario, animacao, jogo, publicidade, outros]
        ano_producao:
          type: integer
        pais_origem:
          type: string
        distribuidora:
          type: string
        sinopse:
          type: string
        duracao_min:
          type: integer
    ConteudoUpdate:
      type: object
      properties:
        titulo:
          type: string
        titulo_original:
          type: string
        tipo:
          type: string
          enum: [filme, serie, documentario, animacao, jogo, publicidade, outros]
        ano_producao:
          type: integer
        pais_origem:
          type: string
        distribuidora:
          type: string
        sinopse:
          type: string
        duracao_min:
          type: integer

    # ─── Descritor ──────────────────────────────────────────────────────────
    Descritor:
      type: object
      required: [id, nome, categoria, descricao]
      properties:
        id:
          type: integer
        nome:
          type: string
        categoria:
          type: string
          enum: [violencia, sexualidade, drogas, linguagem, medo, outros]
        descricao:
          type: string
        icone:
          type: ["string", "null"]
    DescritorInput:
      type: object
      required: [nome, categoria, descricao]
      properties:
        nome:
          type: string
          minLength: 2
        categoria:
          type: string
          enum: [violencia, sexualidade, drogas, linguagem, medo, outros]
        descricao:
          type: string
        icone:
          type: string
    DescritorUpdate:
      type: object
      properties:
        nome:
          type: string
        categoria:
          type: string
          enum: [violencia, sexualidade, drogas, linguagem, medo, outros]
        descricao:
          type: string
        icone:
          type: string

    # ─── Solicitação ────────────────────────────────────────────────────────
...
    Classificacao:
      type: object
      required: [id, conteudo_id, solicitacao_id, faixa_etaria, analista_id, emitida_em]
      properties:
        id:
          type: integer
        conteudo_id:
          type: integer
        conteudo_titulo:
          type: ["string", "null"]
        conteudo_tipo:
          type: ["string", "null"]
        solicitacao_id:
          type: integer
        faixa_etaria:
          type: string
          enum: [L, 10, 12, 14, 16, 18]
        descritores:
          type: array
          items:
            $ref: "#/components/schemas/Descritor"
        analista_id:
          type: integer
        analista_nome:
          type: ["string", "null"]
        justificativa:
          type: ["string", "null"]
        observacoes_publicas:
          type: ["string", "null"]
        numero_certificado:
          type: string
        validade_anos:
          type: ["integer", "null"]
        emitida_em:
          type: string
          format: date-time
        atualizado_em:
          type: string
          format: date-time
    ClassificacaoUpdate:
      type: object
      properties:
        faixa_etaria:
          type: string
          enum: [L, 10, 12, 14, 16, 18]
        descritores_ids:
          type: array
          items:
            type: integer
        justificativa:
          type: string
        observacoes_publicas:
          type: string

    # ─── Dashboard ──────────────────────────────────────────────────────────
...